wip: yieldseeker fix
